+++ src/main/java/io/oreto/brew/data/validation/Size.java
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 @Constraint
 public @interface Size {
-    int value();
+    double value();
     String name() default "";
     String property() default "";
     String group() default "";
